% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_phylo_tree_README.R
\name{build_phylo}
\alias{build_phylo}
\title{build_phylogenetic_tree
This function takes the available trait and distribution data to create
phylogentic topologies. (Not sure if this could be done with what I have, but
it is worth a shot. Perhaps I can do this and/or get tree data from
Open Tree of Life \link{rotl} and do some playing around).}
\usage{
build_phylo(x)
}
\arguments{
\item{x}{name of list 'taxa' (chr \link{1:6}, no default)}
}
\value{
plot_sala_tree the phylogeny built from the 6 genera
}
\description{
build_phylogenetic_tree
This function takes the available trait and distribution data to create
phylogentic topologies. (Not sure if this could be done with what I have, but
it is worth a shot. Perhaps I can do this and/or get tree data from
Open Tree of Life \link{rotl} and do some playing around).
}
