% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_database.R
\name{calc_shrub_vol}
\alias{calc_shrub_vol}
\title{Calculate volume of shrub}
\usage{
calc_shrub_vol(length = 1, width = 1, height = 1)
}
\arguments{
\item{length}{length of the shrub (numeric, default 1)}

\item{width}{width of shrub (numeric, default 1)}

\item{height}{height of shrub (numeric, default 1)}
}
\value{
volume of shrub (numeric)
}
\description{
Calculate volume of shrub
}
\examples{
calc_shrub_vol(2, 3, 2)
calc_shrub_vol(length = 2, width = 3, height = 2)
}
