% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_pop_head_vol_README.R
\name{mean_pop_head_vol}
\alias{mean_pop_head_vol}
\title{calculate mean head volume by population
This function is starting small (really small) and will calculate the mean head volume
for each population in the dataset 'measurements'. The input is the name of the dataset.
The output is an object with the Population and its associated mean head volume (mm).}
\usage{
mean_pop_head_vol(x)
}
\arguments{
\item{x}{name of tibble 'measurements' (tibble, no default)}
}
\value{
Pop_mean_vol which contains Population IDs and associated mean head volume (data frame).
}
\description{
calculate mean head volume by population
This function is starting small (really small) and will calculate the mean head volume
for each population in the dataset 'measurements'. The input is the name of the dataset.
The output is an object with the Population and its associated mean head volume (mm).
}
