% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locality_map_README.R
\name{plot_map}
\alias{plot_map}
\title{plot_locality_map
This function plots sample sites and population data onto a map to visualize population
distributions of Gyrynophilus porphyriticus.}
\usage{
plot_map(x)
}
\description{
plot_locality_map
This function plots sample sites and population data onto a map to visualize population
distributions of Gyrynophilus porphyriticus.
}
