% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_README.R
\name{plot_map}
\alias{plot_map}
\title{plot_map
This function plots sample sites and population data onto a map to visualize population
distributions of Gyrynophilus porphyriticus.}
\usage{
plot_map(x)
}
\arguments{
\item{x}{name of tibble 'measurements' (tibble, no default)}
}
\value{
map_plot1 the plot of the map with populations as red diamonds
}
\description{
plot_map
This function plots sample sites and population data onto a map to visualize population
distributions of Gyrynophilus porphyriticus.
}
